Reset apple music credentials after some duration.
The purpose of this change is to make sure we aren't passing
expired tokens to the server.
We reset credentials if it's been more than a month since user
authenticated with apple music.
This prompts the user to reauthenticate getting fresh tokens.

The monthly thresold is more arbitrary. I suspect that the token
expiry date is longer than a month but I'm just playing it safe.

I also take this opportunity to clean up some redundant logic.